backend/requirements.txt
algorithmia==1.3.0
algorithmia-api-client==1.1.0
asgiref==3.2.7
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.6
django-extensions==2.2.9
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
drf-yasg==1.17.1
enum34==1.1.10
grpcio==1.29.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
packaging==20.4
pdfminer.six==20200517
Pillow==7.1.2
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.7
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.13.0
pytz==2020.1
reportlab==3.5.42
requests==2.23.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sortedcontainers==2.1.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.0.1
zipp==3.1.0
apscheduler
numpy
pandas
sklearn